Refactor the single user form
